refactor(theme): split primitives/colors.js into colors/colors+brand+alpha
Move the single primitives/colors.js into a colors/ folder and extract
the brand (primary, accent, absolute) and alpha (white, black and per-
palette variants) blocks into dedicated modules. Source-of-truth for
each concern now lives in its own file; primitives.brand and
primitives.alpha continue to be composed in colors.js, so existing
tokenRef paths keep resolving.
