+++ packages/theme/src/tokens/primitives/colors/alpha.js
@@ -1,0 +1,116 @@
+export const alpha = {
+  violet: {
+    '50': '#8A84EC0D',
+    '100': '#8A84EC1A',
+    '200': '#8A84EC33',
+    '300': '#8A84EC4D',
+    '400': '#8A84EC66',
+    '500': '#8A84EC80',
+    '600': '#8A84EC99',
+    '700': '#8A84ECB2',
+    '800': '#8A84ECCC',
+    '900': '#8A84ECE5',
+  },
+  orange: {
+    '50': '#FE601F0D',
+    '100': '#FE601F1A',
+    '200': '#FE601F33',
+    '300': '#FE601F4D',
+    '400': '#FE601F66',
+    '500': '#FE601F80',
+    '600': '#FE601F99',
+    '700': '#FE601FB2',
+    '800': '#FE601FCC',
+    '900': '#FE601FE5',
+  },
+  white: {
+    '50': '#FFFFFF0D',
+    '100': '#FFFFFF1A',
+    '200': '#FFFFFF33',
+    '300': '#FFFFFF4D',
+    '400': '#FFFFFF66',
+    '500': '#FFFFFF80',
+    '600': '#FFFFFF99',
+    '700': '#FFFFFFB2',
+    '800': '#FFFFFFCC',
+    '900': '#FFFFFFE5',
+  },
+  black: {
+    '50': '#00000005',
+    '100': '#0000000D',
+    '200': '#00000033',
+    '300': '#0000004D',
+    '400': '#00000066',
+    '500': '#00000080',
+    '600': '#00000099',
+    '700': '#000000B2',
+    '800': '#000000CC',
+    '900': '#000000E5',
+  },
+  green: {
+    '50': '#22C55E0D',
+    '100': '#22C55E1A',
+    '200': '#22C55E33',
+    '300': '#22C55E4D',
+    '400': '#22C55E66',
+    '500': '#22C55E80',
+    '600': '#22C55E99',
+    '700': '#22C55EB2',
+    '800': '#22C55ECC',
+    '900': '#22C55EE5',
+  },
+  red: {
+    '50': '#EF44440D',
+    '100': '#EF44441A',
+    '200': '#EF444433',
+    '300': '#EF44444D',
+    '400': '#EF444466',
+    '500': '#EF444480',
+    '600': '#EF444499',
+    '700': '#EF4444B2',
+    '800': '#EF4444CC',
+    '900': '#EF4444E5',
+  },
+  yellow: {
+    '50': '#F7BD080D',
+    '100': '#F7BD081A',
+    '200': '#F7BD0833',
+    '300': '#F7BD084D',
+    '400': '#F7BD0866',
+    '500': '#F7BD0880',
+    '600': '#F7BD0899',
+    '700': '#F7BD08B2',
+    '800': '#F7BD08CC',
+    '900': '#F7BD08E5',
+  },
+  primary: {
+    '50': '#FE601F0D',
+    '100': '#FE601F1A',
+    '200': '#FE601F33',
+    '300': '#FE601F4D',
+    '400': '#FE601F66',
+    '500': '#FE601F80',
+    '600': '#FE601F99',
+    '700': '#FE601FB2',
+    '800': '#FE601FCC',
+    '900': '#FE601FE5',
+    },
+  accent: {
+    '50': '#8A84EC0D',
+    '100': '#8A84EC1A',
+    '200': '#8A84EC33',
+    '300': '#8A84EC4D',
+    '400': '#8A84EC66',
+    '500': '#8A84EC80',
+    '600': '#8A84EC99',
+    '700': '#8A84ECB2',
+    '800': '#8A84ECCC',
+    '900': '#8A84ECE5',
+    },
+  absolute: {
+    black: '#000000E5',
+    white: '#FFFFFFE5',
+  },
+};
+
+export default { alpha };
