+++ packages/theme/src/tokens/primitives/colors/brand.js
@@ -1,0 +1,40 @@
+export const brand = {
+  primary: {
+    '50': '#FFF5EF',
+    '100': '#FFE7D8',
+    '200': '#FFCFB3',
+    '300': '#FFB180',
+    '400': '#FF8E4D',
+    '500': '#FE601F',
+    '600': '#D94A03',
+    '700': '#B03C02',
+    '800': '#8A2F02',
+    '900': '#692402',
+    '950': '#401602',
+  },
+  accent: {
+    '50': '#F6F6FF',
+    '100': '#ECECFF',
+    '200': '#D9D7FA',
+    '300': '#B5B1F4',
+    '400': '#9F9AF1',
+    '500': '#8A84EC',
+    '600': '#756FE5',
+    '700': '#625BD5',
+    '800': '#524BBB',
+    '900': '#423F99',
+    '950': '#0B0A19',
+  },
+  absolute: {
+    black: '#0A0A0A',
+    white: '#FCFCFC',
+  },
+};
+
+export const brandPrimitives = {
+  primary: brand.primary,
+  accent: brand.accent,
+  absolute: brand.absolute,
+};
+
+export default { brand, brandPrimitives };
