+++ packages/theme/src/tokens/primitives/colors/colors.js
@@ -1,0 +1,155 @@
+import { alpha } from './alpha.js';
+import { brand, brandPrimitives } from './brand.js';
+
+export { brandPrimitives };
+
+export const primitives = {
+  base: {
+    white: '#FFFFFF',
+    black: '#000000',
+    },
+  gray: {
+    '50': '#FCFCFC',
+    '100': '#F2F2F2',
+    '200': '#E5E5E5',
+    '300': '#B2B2B2',
+    '400': '#999999',
+    '500': '#808080',
+    '600': '#666666',
+    '700': '#4D4D4D',
+    '800': '#333333',
+    '900': '#1A1A1A',
+    '950': '#0A0A0A',
+    },
+  // done violet
+  violet: {
+    '50': '#F6F6FF',
+    '100': '#ECECFF',
+    '200': '#D9D7FA',
+    '300': '#B5B1F4',
+    '400': '#9F9AF1',
+    '500': '#8A84EC',
+    '600': '#756FE5',
+    '700': '#625BD5',
+    '800': '#524BBB',
+    '900': '#423F99',
+    '950': '#0B0A19',
+    },
+  orange: {
+    '50': '#FFF5EF',
+    '100': '#FFE7D8',
+    '200': '#FFCFB3',
+    '300': '#FFB180',
+    '400': '#FF8E4D',
+    '500': '#FE601F',
+    '600': '#D94A03',
+    '700': '#B03C02',
+    '800': '#8A2F02',
+    '900': '#692402',
+    '950': '#401602',
+    },
+  slate: {
+    '50': '#F8F7FC',
+    '100': '#F0EFF8',
+    '200': '#E3E1F0',
+    '300': '#D3D2E6',
+    '400': '#BEBBD6',
+    '500': '#A39FBE',
+    '600': '#8884A4',
+    '700': '#6D698B',
+    '800': '#534F70',
+    '900': '#353040',
+    '950': '#13131A',
+    },
+  yellow: {
+    '50': '#FEF8E6',
+    '100': '#FDF2CE',
+    '200': '#FCE49C',
+    '300': '#FAD76B',
+    '400': '#F8CA3A',
+    '500': '#F7BD08',
+    '600': '#C59707',
+    '700': '#947105',
+    '800': '#634B03',
+    '900': '#312602',
+    '950': '#191301',
+    },
+  green: {
+    '50': '#F0FDF4',
+    '100': '#D0FBDF',
+    '200': '#A4F4C0',
+    '300': '#78EDA2',
+    '400': '#52E086',
+    '500': '#26D968',
+    '600': '#18B452',
+    '700': '#16833F',
+    '800': '#12542B',
+    '900': '#0A2916',
+    '950': '#03170B',
+    },
+  blue: {
+    '50': '#EFF6FF',
+    '100': '#DBEAFE',
+    '200': '#BFDBFE',
+    '300': '#93C5FD',
+    '400': '#60A5FA',
+    '500': '#3B82F6',
+    '600': '#2563EB',
+    '700': '#1D4ED8',
+    '800': '#1E40AF',
+    '900': '#1E3A8A',
+    '950': '#172554',
+    },
+  neutral: {
+    '50': '#FAFAFA',
+    '100': '#F5F5F5',
+    '200': '#E5E5E5',
+    '300': '#D4D4D4',
+    '400': '#A3A3A3',
+    '500': '#737373',
+    '600': '#525252',
+    '700': '#404040',
+    '800': '#262626',
+    '900': '#171717',
+    '950': '#0A0A0A',
+    },
+  red: {
+    '50': '#FFE5E5',
+    '100': '#FFCCCC',
+    '200': '#FECACA',
+    '300': '#FCA5A5',
+    '400': '#F87171',
+    '500': '#EF4444',
+    '600': '#DC2626',
+    '700': '#B91C1C',
+    '800': '#991B1B',
+    '900': '#7F1D1D',
+    '950': '#240505',
+    },
+  brand,
+  surface: {
+    0: '#FFFFFF',
+    50: '#FAFAFA',
+    100: '#F5F5F5',
+    200: '#E5E5E5',
+    300: '#D4D4D4',
+    400: '#A3A3A3',
+    500: '#737373',
+    600: '#525252',
+    700: '#404040',
+    800: '#262626',
+    900: '#171717',
+    950: '#0A0A0A',
+  },
+  alpha,
+};
+
+export const surfacePrimitives = {
+  surface: primitives.surface,
+};
+
+export default {
+  primitives,
+  surfacePrimitives,
+  brandPrimitives,
+};
